move create_detector and DetectorTypeEnum
