--- frigate/detectors/detector_type.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-
-class DetectorTypeEnum(str, Enum):
-    edgetpu = "edgetpu"
-    openvino = "openvino"
-    cpu = "cpu"
-    tensorrt = "tensorrt"
